fix: Support Pytest 7 and run with importlib
