+++ services/datalad/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+pythonpath = .
+addopts = "--import-mode=importlib"
