+++ services/datalad/tests/test_auth.py
@@ -2,7 +2,6 @@ import base64
 from falcon import Request, Response
 from datalad_service.middleware.auth import parse_authorization_header, AuthenticateMiddleware
 
-from .dataset_fixtures import *
 
 raw_token = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJmZDQ0ZjVjNS1iMjFiLTQyMGItOTU1NS1hZjg1NmVmYzk0NTIiLCJlbWFpbCI6Im5lbGxAc3F1aXNoeW1lZGlhLmNvbSIsInByb3ZpZGVyIjoiZ29vZ2xlIiwibmFtZSI6Ik5lbGwgSGFyZGNhc3RsZSIsImFkbWluIjp0cnVlLCJzY29wZXMiOlsiZGF0YXNldDpnaXQiXSwiZGF0YXNldCI6ImRzMDAyNDA0IiwiaWF0IjoxNjA4NDEwNjEyLCJleHAiOjIxNDc0ODM2NDd9.Gxy-2weKFlQ-q1IcP3X0OyqL4uF3WIXtR-ntRsUB3Gk'
 basic_token = base64.urlsafe_b64encode(
