+++ services/datalad/tests/test_draft.py
@@ -1,8 +1,8 @@
-import falcon
-import jwt
 import os
 
-from .dataset_fixtures import *
+import falcon
+import json
+import jwt
 
 
 def test_add_commit_info(client):
