+++ services/datalad/tests/test_publish.py
@@ -1,5 +1,5 @@
+import json
 import os
-from .dataset_fixtures import *
 from unittest.mock import Mock, call
 
 import falcon
